Permit variable size >= 100 MB
1. Make the Fortran variable an integer(kind=8)
2. Use an i12 format to print out the value, not i8

+++ frame/module_alloc_space.h
@@ -53,7 +53,7 @@
       INTEGER ierr
 
       INTEGER                              :: loop
-      INTEGER                              :: nba ! number of bytes allocated per variable
+      INTEGER(KIND=8)                      :: nba ! number of bytes allocated per variable
       CHARACTER(LEN=256)                   :: message_string
 
    ! Local data
